Move into util file
