--- src/lib/util/pluraliseItemName.ts
@@ -1,3 +1,0 @@
-export function pluraliseItemName(name: string): string {
-	return name + (name.endsWith('s') ? '' : 's');
-}
