+++ src/mahoji/lib/abstracted_commands/buryCommand.ts
@@ -7,7 +7,7 @@ import { BuryingActivityTaskOptions } from '../../../lib/types/minions';
 import { formatDuration, stringMatches } from '../../../lib/util';
 import addSubTaskToActivityTask from '../../../lib/util/addSubTaskToActivityTask';
 import { calcMaxTripLength } from '../../../lib/util/calcMaxTripLength';
-import { pluraliseItemName } from '../../../lib/util/pluraliseItemName';
+import { pluraliseItemName } from '../../../lib/util/smallUtils';
 
 export async function buryCommand(user: MUser, channelID: string, boneName: string, quantity?: number) {
 	const speedMod = 1;
