+++ src/mahoji/lib/abstracted_commands/castCommand.ts
@@ -8,7 +8,7 @@ import { formatDuration, stringMatches } from '../../../lib/util';
 import addSubTaskToActivityTask from '../../../lib/util/addSubTaskToActivityTask';
 import { calcMaxTripLength } from '../../../lib/util/calcMaxTripLength';
 import { determineRunes } from '../../../lib/util/determineRunes';
-import { pluraliseItemName } from '../../../lib/util/pluraliseItemName';
+import { pluraliseItemName } from '../../../lib/util/smallUtils';
 import { updateBankSetting } from '../../../lib/util/updateBankSetting';
 import { userHasGracefulEquipped } from '../../mahojiSettings';
