+++ src/mahoji/lib/abstracted_commands/enchantCommand.ts
@@ -7,7 +7,7 @@ import { formatDuration, itemNameFromID, stringMatches } from '../../../lib/util
 import addSubTaskToActivityTask from '../../../lib/util/addSubTaskToActivityTask';
 import { calcMaxTripLength } from '../../../lib/util/calcMaxTripLength';
 import { determineRunes } from '../../../lib/util/determineRunes';
-import { pluraliseItemName } from '../../../lib/util/pluraliseItemName';
+import { pluraliseItemName } from '../../../lib/util/smallUtils';
 import { updateBankSetting } from '../../../lib/util/updateBankSetting';
 
 export async function enchantCommand(user: MUser, channelID: string, name: string, quantity?: number) {
